project/requirements_dev.txt
nose==1.3.7
mysql-connector==2.1.6